quick fix for amount data type and allproducts own by a user

+++ database/migrations/2023_01_22_053149_create_products_table.php
@@ -21,7 +21,7 @@ class CreateProductsTable extends Migration
             $table->string('description')->nullable();
             $table->integer('quantity')->nullable();
             $table->integer('unit_price')->nullable();
-            $table->bigInteger('amount');
+            $table->decimal('amount', 5, 2);
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
